+++ libresin/libresin/core/sdf_tree/group_node.hpp
@@ -2,6 +2,7 @@
 #define RESIN_GROUP_NODE_HPP
 #include <libresin/core/sdf_shader_consts.hpp>
 #include <libresin/core/sdf_tree/sdf_tree.hpp>
+#include <libresin/core/sdf_tree/sdf_tree_node.hpp>
 #include <libresin/core/transform.hpp>
 #include <libresin/utils/logger.hpp>
 #include <memory>
