+++ libresin/libresin/core/sdf_tree/primitive_node.hpp
@@ -1,9 +1,10 @@
 #ifndef RESIN_PRIMITIVE_NODE_HPP
 #define RESIN_PRIMITIVE_NODE_HPP
 #include <format>
-#include <libresin/core/component_id_registry.hpp>
+#include <libresin/core/id_registry.hpp>
 #include <libresin/core/sdf_shader_consts.hpp>
 #include <libresin/core/sdf_tree/sdf_tree.hpp>
+#include <libresin/core/sdf_tree/sdf_tree_node.hpp>
 #include <libresin/core/transform.hpp>
 
 namespace resin {
